+++ CHANGES.md
@@ -9,7 +9,8 @@
 - FIX: Fix issue with geocoding with unzipped `Sentinel-3 OLCI` product ([#137](https://github.com/sertit/eoreader/issues/137))
 - FIX: Fix iceye product when extent file (*.kml) not found ([#135](https://github.com/sertit/eoreader/pull/135))
 - FIX: Add missing `pystac[validation]` in setup.py
-- CI: Fix S3 endpoint management with `sertit==1.37.0`
+- FIX: Handle RCM and RS2 products that doesn't bundle their extent in a KML file ([#155](https://github.com/sertit/eoreader/issues/155))
+- CI: Fix S3 endpoint management with `sertit==1.37`
 
 ## 0.21.1 (2024-04-03)